Feature: re-sign all active waves on TUF targets key rotation (#352)
* Feature: re-sign all active waves on TUF targets key rotation
* Feature: ability to manually re-sign wave targets on TUF root updates
* Feature: handy ability to show only active waves list
* Cleanup: use same type for limit and page in wave and device lists
  Both arguments must be positive integers, but we had different logic:
* Improve: fail sign-prod-targets for non-existing tags

+++ subcommands/common.go
@@ -269,7 +269,7 @@ func FindWritableDirInPath(helperPath string) string {
 	return ""
 }
 
-func ShowPages(showPage int, next *string) {
+func ShowPages(showPage uint64, next *string) {
 	if next != nil {
 		fmt.Print("\nNext page can be viewed with: ")
 		found := false
